chore: use `GPL-3.0` as this project's LICENSE
v0.0.0 was published with `GPL-2.0` to be conservative, but that changes
right now!

GPLv3 looks like the way forward WRT community. This was discussed with
`vcsh` upstream in https://github.com/RichiH/vcsh/issues/317. This
project's concept of overlay repositories (potentially overlapping repo
work trees rooted at the home directory of a user) are basically design
drawn from the `vcsh` project, so credit (and licensing) where credit is
due!

Also, start a manual release checklist for the project, since I don't
know if we'll ever be able to fully automate anything. Definitely not at
this point!

Resolves #9.